feat(#128): 클래스 상세 조회 변경된 API 수정
